Extract roads out
