+++ lib/index.ts
@@ -158,8 +158,14 @@ export * from './src/utils'
 export * from './src/weather'
 export * from './src/weightedRandomFetcher'
 export * from './src/weightRandom'
-
 export * from './src/worldType'
+
+export * from './roads/roads'
+export * from './roads/RoadType'
+export * from './roads/roadTypes'
+export * from './roads/assignRoad'
+export * from './roads/findExistingRoad'
+
 export * from './tavern/_common'
 export * from './tavern/createTavernName'
 export * from './tavern/getTavernSin'
@@ -184,7 +190,6 @@ export * from './town/getTownType'
 export * from './town/getTownWelfare'
 export * from './town/politicsWeightedRoll'
 export * from './town/replaceTownName'
-export * from './town/roads'
 export * from './town/setBaseDemographics'
 export * from './town/setTownMaterialProbability'
 export * from './town/townData'
