+++ lib/roads/RoadType.ts
@@ -1,0 +1,18 @@
+
+export type RoadType = 'street' |
+  'lane' |
+  'road' |
+  'square' |
+  'way' |
+  'crescent' |
+  'close' |
+  'wynd' |
+  'row' |
+  'dyke' |
+  'avenue' |
+  'alley' |
+  'drive' |
+  'boulevard' |
+  'plaza' |
+  'track' |
+  'trail';
